Refactor Ollama provider tests into focused modules
